Extract vscalar tests
Keep callback coverage intact while leaving room for native callback tests in
the production module.
